Change the agent usage data to floating point (#73)

+++ packages/narada-core/src/narada_core/actions/models.py
@@ -22,7 +22,7 @@ _MaybeStructuredOutput = TypeVar("_MaybeStructuredOutput", bound=BaseModel | Non
 
 class AgentUsage(BaseModel):
     actions: int
-    credits: int
+    credits: float
 
 
 class OperatorActionTraceItem(BaseModel):
